raise go directive to 1.26.4 to clear stdlib CVEs; gosec annotations; commit go.sum
govulncheck flagged six stdlib vulnerabilities reachable from connector
HTTP and the ratelimit transport that are only fixed in 1.26.4. Bumped
the go.mod directive and committed the regenerated go.sum so CI no
longer needs go mod tidy.

gosec G204 on the git wrapper and G115 on the PR-number int32 casts
are annotated with their justifications. PR numbers fit in int32; the
git binary path and argv are internally sourced.

+++ internal/connectors/github/pr_meta.go
@@ -60,9 +60,10 @@ func (c *Connector) extractPRReviewRequests(ctx context.Context, repo connector.
 			return
 		}
 		vars := map[string]any{
-			"owner":  githubv4.String(owner),
-			"name":   githubv4.String(name),
-			"number": githubv4.Int(number),
+			"owner": githubv4.String(owner),
+			"name":  githubv4.String(name),
+			// #nosec G115 -- PR numbers fit comfortably in int32.
+			"number": githubv4.Int(int32(number)),
 			"after":  cursor,
 		}
 		if err := c.gql.Query(ctx, &q, vars); err != nil {
