+++ internal/connectors/github/prs.go
@@ -384,9 +384,10 @@ func (c *Connector) paginatePRCommits(ctx context.Context, owner, name string, n
 			} `graphql:"repository(owner: $owner, name: $name)"`
 		}
 		vars := map[string]any{
-			"owner":  githubv4.String(owner),
-			"name":   githubv4.String(name),
-			"number": githubv4.Int(number),
+			"owner": githubv4.String(owner),
+			"name":  githubv4.String(name),
+			// #nosec G115 -- PR numbers fit comfortably in int32.
+			"number": githubv4.Int(int32(number)),
 			"after":  githubv4.String(cursor),
 		}
 		if err := c.gql.Query(ctx, &q, vars); err != nil {
